inital game&snake separation rebase merge conflicts
