+++ pyClient.c
@@ -1,0 +1,20 @@
+#include "pyClient.h"
+
+int createPod() {
+    wchar_t *program = Py_DecodeLocale(argv[0], NULL);
+    if (program == NULL) {
+        fprintf(stderr, "Fatal error: cannot decode argv[0]\n");
+        exit(1);
+    }
+    Py_SetProgramName(program);  /* optional but recommended */
+    Py_Initialize();
+
+    FILE *fp = fopen("createPod.py","rb");
+    PyRun_SimpleFile(fp,"createPod.py");   
+
+    if (Py_FinalizeEx() < 0) {
+        exit(120);
+    }
+    PyMem_RawFree(program);
+    return 0;
+}
